+++ test.cpp
@@ -17,8 +17,8 @@ void testFormatter(){
   assert(format.kNumCols_ == 3);
   assert(format.kNumRows_ == 3);
   
-  std::vector<std::vector<int32_t>> formatted_matrix_1 = format.formattedMatrix_;
-  std::vector<std::vector<int32_t>> correct_matrix_1 
+  std::vector<std::vector<float>> formatted_matrix_1 = format.formattedMatrix_;
+  std::vector<std::vector<float>> correct_matrix_1 
   = {{7,7,7,0,0,0},{1,2,3,1,0,0},{4,5,6,0,1,0},{7,8,9,0,0,1}};
   
   for(size_t i = 0; i < formatted_matrix_1.size(); ++i){
@@ -27,8 +27,8 @@ void testFormatter(){
     }
   }
 
-  std::vector<int32_t> formatted_vector_1 = format.formattedVector_;
-  std::vector<int32_t> correct_vector_1 = {10,10,11};
+  std::vector<float> formatted_vector_1 = format.formattedVector_;
+  std::vector<float> correct_vector_1 = {10,10,11};
 
   for (size_t i = 0; i < formatted_vector_1.size(); ++i){
     assert(formatted_vector_1.at(i) == correct_vector_1.at(i));
@@ -43,8 +43,8 @@ void testFormatter(){
   assert(format.kNumCols_ == 2);
   assert(format.kNumRows_ == 2);
   
-  std::vector<std::vector<int32_t>> formatted_matrix_2 = format.formattedMatrix_;
-  std::vector<std::vector<int32_t>> correct_matrix_2 = {{-1,-10,0,0},{1,20,1,0},{0,1,0,1}};
+  std::vector<std::vector<float>> formatted_matrix_2 = format.formattedMatrix_;
+  std::vector<std::vector<float>> correct_matrix_2 = {{-1,-10,0,0},{1,20,1,0},{0,1,0,1}};
   
   for(size_t i = 0; i < formatted_matrix_2.size(); ++i){
     for (size_t j = 0; j < formatted_matrix_2.at(0).size(); ++j){
@@ -52,8 +52,8 @@ void testFormatter(){
     }
   }
 
-  std::vector<int32_t> formatted_vector_2 = format.formattedVector_;
-  std::vector<int32_t> correct_vector_2 = {100,1};
+  std::vector<float> formatted_vector_2 = format.formattedVector_;
+  std::vector<float> correct_vector_2 = {100,1};
 
   for (size_t i = 0; i < formatted_vector_2.size(); ++i){
     assert(formatted_vector_2.at(i) == correct_vector_2.at(i));
